+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 16.0.0
+## 16.0.0 (January 29, 2018)
 
 ### Enhancements
 - **General Changes**